Set up code style auto-fix (#84)
This sets up GitHub to automatically fix any 'code style issues'
(whitespace, etc) using the checks @paulrobertlloyd set up in #79.

This will be extra useful for anyone making commits and opening pull
requests using the GitHub web interface, who might not be able to easily
run the code style checks locally.


It does it by running `npm run lint:fix` and then committing and pushing
any changes if there are any.

One wrinkle is that any commits pushed using the default `GITHUB_TOKEN`
available to workflows do no themselves then trigger any further
workflow runs, to avoid accidental recursion. (See
[guidance](https://docs.github.com/en/actions/how-tos/write-workflows/choose-when-workflows-run/trigger-a-workflow)).
To avoid this, the workflow is set up to use a token added to the
secrets as `AUTOLINT_TOKEN`, with permissions for this repository only.

The token will expire after a year so we’d have to remember to create a
new one once a year.